Add mono to project (#13)
Use our own tooling for development in the Elixir integration repos.
https://github.com/appsignal/mono/
Follow the instructions on the mono README to install it locally.

Mono repo support for Elixir isn't 100% feature complete yet so we're
first adding it to every repo separately. Right now we'll use it for
changesets, changelogs and publishing the package.

I didn't add mono to the Semaphore CI test setup as it doesn't really
add anything there at this point. I also didn't add it to the main
Elixir package's CI yet either.