Refactor Git lint script
Combine everything in one script so it's easier to maintain, the env var
with the version, the installation and running the command.

If the env vars are in the `semaphore.yml` file, it would also need to
update the `semaphore.yml` file every time you change it. It's easy to
forget, and would need multiple builds to fix.