Source `bin/setup` instead of running it
The Semaphore-provided utilities, such as `sem-version`, are
provided as shell functions. To use them from a script, the script
must be sourced, not invoked. As such, before this commit, calls
to `sem-version` within the `bin/setup` script were failing, and
all the test runs were using the versions of Elixir and Erlang
provided by Semaphore by default.

This commit also renames the Elixir `master` version to `main`,
which was not failing before for the same reason.