Add /already-sent regression test
As reported in
https://github.com/appsignal/appsignal-elixir-plug/issues/26, raising an
error after sending the response causes `Appsignal.Plug.handle_error/5`
to crash, raising an `AlreadySentError` because it calls
`Plug.Conn.put_status/2`.

This patch adds a passing test with a Plug endpoint like this;

    get "/already-sent" do
      send_resp(conn, 200, "Exception!")
      raise "Exception!"
    end

This test doesn't fail because the `conn` used while handling the error
isn't the one returned from the call to `Plug.Conn.send_resp/3`.