Ensure spans with errors aren't double-closed (#51)
Since
https://github.com/appsignal/appsignal-elixir/commit/d218b4054f7c5accb94d07d215817ddc9d0b3a3c,
the instrument/1 function automatically closes the created span. This
caused issues with the Plug integration, which assumed spans weren't
closed.

This patch moves off instrument/1 in the Plug integration by starting
and closing the span itself. Because it now opens the span itself in
the "http_request" namespace, it no longer has to set the namespace.