Remove unnecessary Plug.ErrorHandler from Plug (#55)
I don't think we need this. Our own error handler logic, the try-catch
in the `call` function does all the error handling we need.

By including it, we change the behavior of errors that occur in
requests, because the `Plug.ErrorHandler`'s default implementation makes
it return a response with status 500 and a body that says "Something
went wrong".

Source:
https://github.com/elixir-plug/plug/blob/1947edc8171f923ecc262efba6a0946264fa5582/lib/plug/error_handler.ex#L66

Closes #54